+++ modules/template/elements/ContainerDefinition.php
@@ -21,7 +21,7 @@ use Yii;
  *
  * @property-read Template[] $allowedTemplates
  */
-class ContainerDefinition extends BaseTemplateElementContentDefinition
+class ContainerDefinition extends BaseTemplateElementDefinition
 {
     /**
      * @var Template[] Cached templates selected as allowed for this definition
@@ -92,5 +92,13 @@ class ContainerDefinition extends BaseTemplateElementContentDefinition
     public function isSingleAllowedTemplate(): bool
     {
         return is_array($this->templates) && count($this->templates) === 1;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public static function find()
+    {
+        return parent::find()->where([self::tableName() . '.content_type' => ContainerElement::class]);
     }
 }