+++ docs/CHANGELOG.md
@@ -13,6 +13,7 @@ Changelog
 - Enh #367: Migrate all element contents to single table
 - Fix #376: Use template name for allowed templates in the container element dynamic attributes
 - Fix #377: Rename "snipped" to "snippet"
+- Enh #378: Refactor template instance and owner content
 
 1.10.11 (Unreleased)
 --------------------------